+++ package.json
@@ -17,7 +17,6 @@
     "react": "^18",
     "react-dom": "^18",
     "react-hook-form": "^7.53.0",
-    "react-router-dom": "^6.27.0",
     "yup": "^1.4.0"
   },
   "devDependencies": {
