+++ src/Utlis/useInputRegister.tsx
@@ -1,0 +1,15 @@
+import { yupResolver } from '@hookform/resolvers/yup';
+import { validationSchema } from './Validation';
+import { useForm } from 'react-hook-form';
+
+
+const useInputRegister = () => {
+
+  const formObject = useForm<{ taskName: string }>({
+    resolver: yupResolver(validationSchema),
+});
+
+  return { formObject };
+};
+
+export default useInputRegister;
