+++ src/app/msw-provider.tsx
@@ -13,7 +13,7 @@ const mockingEnabledPromise =
               {
               return;
             }
-            print.warning();
+            // print.warning();
           },
         });
       })
